Cache all achievements in SavedVariables
GetAchievementInfo is really slow so this is pretty much the only
option. The cache will reset on new ToC versions. If we need to modify
the cache in the future, we can clear it if there's no addon version
field and add one going forward.

The cache could be distributed with the addon, but then updates will be
required when new achievements are added.

I tried using LibSerialize and AceSerializer to serialize the cache, but
both take 300+ ms for me. That plus the 150 for compression using
LibDeflate would make this by far the slowest provider.