Determine LimitedTooltip's whitelist automatically
If we first filter to only functions that GameTooltip has, but not its
parent, Frame, that leaves us with pretty close to what we want. Then
we can filter out everything except setters and manually add in
any setters that don't start with Set.

It's important for this to not be a manual process so that new methods
added in future expansions aren't excluded.