Revert "ToysSearchProvider: fix toys failing to load under high load"
This reverts commit a2aa06750e68af7f3bcff7c5377fafde8ba067d3.

The reverted commit was ineffective in fixing the issue.

I did some more investigation and found the mistakes in my previous
approaches.

The first attempt refers to the state after reverting this commit. The
second attempt refers to the state before reverting this commit.

My first attempt used a throttle on TOYS_UPDATED since I was under
the impression that the event would fire a maximum of two times, and
both events would be in quick succession. Turns out that's false,
TOYS_UPDATED fires in rapid succession from the point that the toy
filter is changed until the point that all toys are loaded. If you
spam transmog functions, this event will keep spamming until the
transmog functions stop.

My second attempt was more focused on toy names being nil, so I
essentially just kept checking every second if any of the toy names are
nil. If not, and the number of toys available is the same as the number
of toys available a second ago, we are done. Turns out it will quickly
load around 30-60% of the toys depending on how hard I spam transmog
functions, return non-nil names for those toys, and then stop loading
any more toys until the transmog functions stop. Despite not loading
any more toys for this period of time, TOYS_UPDATED is still repeatedly
fired.

Looks like all I need to do is go back to something like the first
solution, but with a debounce rather than throttle.