ToysSearchProvider: fix waiting too long and toy names becomnig nil again
Toy names are only guaranteed to not be nil exactly when the last
TOYS_UPDATED fires. The debounce was retrieving items a second later
than that, so toys were nil again. The solution is to retrieve the
latest items on every TOYS_UPDATED, but not set them via
GlobalSearch_ProviderItemsUpdated message until the debounce.