Fix dragging search results
Due to actions triggering on mouse down, it was impossible to drag
search results (ex. drag spells onto bars).

I attempted to fix it a while back:
1af4140c1d81b6f245ca4bbfa20aac66698fea45
but reverted it due to the fix breaking all mouse actions on retail:
1af4140c1d81b6f245ca4bbfa20aac66698fea45

This reintroduces the classic fix as well as a workaround for retail.